[MIG] sign_oca: Migration to 16.0
[UPD] Update sign_oca.pot
