+++ sign_oca/__manifest__.py
@@ -5,7 +5,7 @@
     "name": "Sign Oca",
     "summary": """
         Allow to sign documents inside Odoo CE""",
-    "version": "14.0.1.0.1",
+    "version": "16.0.1.0.1",
     "license": "AGPL-3",
     "author": "Dixmit,Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/sign",
@@ -26,12 +26,40 @@
     "demo": [
         "demo/sign_oca_template.xml",
     ],
-    "qweb": [
-        "static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.xml",
-        "static/src/components/sign_oca_configure/sign_oca_configure.xml",
-        "static/src/components/sign_oca_pdf/sign_oca_pdf.xml",
-        "static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.xml",
-        "static/src/elements/elements.xml",
-    ],
+    "assets": {
+        "web.assets_backend": [
+            "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.xml",
+            "sign_oca/static/src/components/sign_oca_configure/sign_oca_configure.xml",
+            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.xml",
+            "sign_oca/static/src/elements/elements.xml",
+            "sign_oca/static/src/scss/sign_oca.scss",
+            "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.esm.js",
+            "sign_oca/static/src/components/sign_oca_configure/sign_oca_configure.esm.js",
+            "sign_oca/static/src/elements/text.esm.js",
+            "sign_oca/static/src/elements/signature.esm.js",
+            "sign_oca/static/src/elements/check.esm.js",
+            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.esm.js",
+            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf_action.esm.js",
+        ],
+        "web.assets_frontend": [
+            "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.xml",
+            "sign_oca/static/src/components/sign_oca_configure/sign_oca_configure.xml",
+            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.xml",
+            "sign_oca/static/src/elements/elements.xml",
+            "sign_oca/static/src/scss/sign_oca.scss",
+            "sign_oca/static/src/components/sign_oca_pdf_common/sign_oca_pdf_common.esm.js",
+            "sign_oca/static/src/elements/text.esm.js",
+            "sign_oca/static/src/elements/signature.esm.js",
+            "sign_oca/static/src/elements/check.esm.js",
+            "sign_oca/static/src/components/sign_oca_pdf/sign_oca_pdf.esm.js",
+            "sign_oca/static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.esm.js",
+            "sign_oca/static/src/components/sign_oca_pdf_portal/sign_oca_pdf_portal.xml",
+            "sign_oca/static/src/scss/portal.scss",
+        ],
+        "sign_oca.sign_assets": [
+            "sign_oca/static/src/scss/sign.scss",
+            "web/static/src/libs/fontawesome/css/font-awesome.css",
+        ],
+    },
     "maintainers": ["etobella"],
 }
