+++ sign_oca/controllers/main.py
@@ -12,22 +12,17 @@ from odoo.addons.portal.controllers.portal import CustomerPortal
 class SignController(http.Controller):
     @http.route("/sign_oca/get_assets.<any(css,js):ext>", type="http", auth="public")
     def get_sign_resources(self, ext):
-        xmlid = "sign_oca.sign_assets"
-        files, _remains = request.env["ir.qweb"]._get_asset_content(
-            xmlid, options=request.context
-        )
-        asset = AssetsBundle(xmlid, files)
+        bundle = "sign_oca.sign_assets"
+        files, _ = request.env["ir.qweb"]._get_asset_content(bundle)
+        asset = AssetsBundle(bundle, files)
         mock_attachment = getattr(asset, ext)()
         if isinstance(
             mock_attachment, list
         ):  # suppose that CSS asset will not required to be split in pages
             mock_attachment = mock_attachment[0]
-        _status, headers, content = request.env["ir.http"].binary_content(
-            id=mock_attachment.id, unique=asset.checksum
-        )
-        content_base64 = base64.b64decode(content) if content else ""
-        headers.append(("Content-Length", len(content_base64)))
-        return request.make_response(content_base64, headers)
+        stream = request.env["ir.binary"]._get_stream_from(mock_attachment)
+        response = stream.get_response()
+        return response
 
 
 class PortalSign(CustomerPortal):
@@ -76,7 +71,7 @@ class PortalSign(CustomerPortal):
         except (AccessError, MissingError):
             return request.redirect("/my")
         data = io.BytesIO(base64.standard_b64decode(signer_sudo.request_id.data))
-        return http.send_file(data)
+        return http.send_file(data, filename=signer_sudo.request_id.name)
 
     @http.route(
         ["/sign_oca/info/<int:signer_id>/<string:access_token>"],
