+++ sign_oca/static/src/components/sign_oca_configure/sign_oca_configure.xml
@@ -37,12 +37,15 @@
                 title="Signature configuration"
             />
             <div class="o_sign_oca_resize fa fa-arrows-alt" style="z-index: 90;" />
-            <span class="o_sign_placeholder" t-esc="name" />
+            <span class="o_sign_placeholder" t-out="name" />
 
         </div>
     </t>
     <t t-name="sign_oca.sign_iframe_contextmenu">
-        <div class="o_sign_oca_context_menu" t-attf-style="top: #{top};left: #{left}">
+        <div
+            class="o_sign_oca_context_menu"
+            t-attf-style="top: #{top};left: #{left};position: absolute"
+        >
             <div
                 class="o_sign_oca_button_header"
             >Click on the field that you want to add</div>
@@ -50,7 +53,7 @@
                 <div
                     class="o_sign_oca_context_menu_item"
                     t-att-data-field="field.id"
-                    t-esc="field.name"
+                    t-out="field.name"
                     t-att-data-page="page_id"
                 />
             </t>
@@ -64,7 +67,7 @@
                     <t t-foreach="info.fields" t-as="field">
                         <option
                             t-att-value="field.id"
-                            t-esc="field.name"
+                            t-out="field.name"
                             t-att-selected="item.field_id === field.id ? 'selected': None"
                         />
                     </t>
@@ -76,7 +79,7 @@
                     <t t-foreach="info.roles" t-as="role">
                         <option
                             t-att-value="role.id"
-                            t-esc="role.name"
+                            t-out="role.name"
                             t-att-selected="item.role_id === role.id ? 'selected': None"
                         />
                     </t>
