--- sign_oca/static/src/elements/registry.js
@@ -1,5 +1,0 @@
-odoo.define("sign_oca.SignRegistry", function (require) {
-    "use strict";
-    const Registry = require("web.Registry");
-    return new Registry();
-});
