+++ sign_oca/tests/test_sign.py
@@ -4,10 +4,10 @@
 import base64
 
 from odoo.modules.module import get_module_resource
-from odoo.tests.common import Form, SavepointCase
+from odoo.tests.common import Form, TransactionCase
 
 
-class TestSign(SavepointCase):
+class TestSign(TransactionCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
